# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: Nikah Time
  license:
    name: MIT
  version: 0.1.1
servers:
  - url: /api/
paths:
...
  /chats/add/{userId}:
    post:
      tags:
        - Chats
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: Айдишник пользователя, с которым авторизованный пользователь начинает чат
      summary: Создание нового чата
      operationId: addUserChat
      responses:
        "200":
          description: Чат создан успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserChat'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /chats/get/{chatId}:
    get:
      tags:
        - Chats
      parameters:
        - in: path
          name: chatId
          schema:
            type: integer
          required: true
          description: Айдишник чата, сообщения которого нужно получить
      summary: Получение информации о чате между пользователями
      operationId: getChatMessages
      responses:
        "200":
          description: Информация о чате между пользователями
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Messages'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /chats/block/{chatId}:
    put:
      tags:
        - Chats
      parameters:
        - in: path
          name: chatId
          schema:
            type: integer
          required: true
          description: Айдишник чата, который нужно заблокировать/разблокировать
      summary: Заблокировать/разблокировать чат между пользователями
      operationId: blockChat
      responses:
        "200":
          description: Чат заблокирован/разблокирован успешно. Обновляем данные чата пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Messages'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /chats/send/message:
    post:
      tags:
        - Chats
      summary: Отправка сообщения
      operationId: sendMessage
      requestBody:
        description: Данные о сообщении
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessage'
            examples:
              default:
                value:
                  message: Hello
                  chatId: 1
                  messageType: text
        required: true
      responses:
        "200":
          description: Сообщение отправлено успешно. После этого будет вызываться событие об уведомлении пользователя, которому отправлено сообщение
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /chats/delete/{chatId}:
    delete:
      tags:
        - Chats
      parameters:
        - in: path
          name: chatId
          schema:
            type: integer
          required: true
          description: Айдишник чата, который нужно удалить
      summary: Удалить чат между пользователями
      operationId: deleteChat
      responses:
        "200":
          description: Чат удален успешно. Обновляем данные о чатах пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chats'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /chats/get/message/{messageId}:
    get:
      tags:
        - Chats
      parameters:
        - in: path
          name: messageId
          schema:
            type: integer
          required: true
          description: Айдишник сообщения, данные которого нужно получить
      summary: Получение информации о сообщении пользователя
      operationId: getMessage
      responses:
        "200":
          description: Информация о сообщении между пользователями
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /chats/seen/message/{messageId}:
    put:
      tags:
        - Chats
      parameters:
        - in: path
          name: messageId
          schema:
            type: integer
          required: true
          description: Айдишник сообщения, которое было прочитано
      summary: Отметка сообщения прочитанным
      operationId: makeSeenMessage
      responses:
        "200":
          description: Сообщение отмечено прочитанным
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
...
  /store/file:
    post:
      tags:
        - Account
      summary: Загрузка файлов на сервер
      operationId: storeFile
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/File'
      responses:
        "200":
          description: Загрузка файла прошла успешно
          content:
            application/json:
              schema:
                properties:
                  fileUrl:
                    type: string
                    description: 'Ссылка на файл на сервере'
                  fileType:
                    type: string
                    enum:
                      - image
                      - file
                      - video
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    UserChat:
      type: object
      properties:
        chatId:
          type: integer
          format: int64
          description: id созданного чата
        userAvatar:
          type: string
          description: Аватар пользователя, с которым создан чат
        userName:
          type: string
          description: Имя пользователя, с которым создан чат
        isOnline:
          type: boolean
...
    File:
      required:
        - fileType
        - file
      type: object
      properties:
        fileType:
          type: string
          enum:
            - image
            - file
            - video
        file:
          type: string
          format: binary
          description: "Доступные расширения картинок - jpg,png,jpeg,bmp. На файлы нет ограничений. На видео - не более 100мб"
...
    Chats:
      type: array
      items:
        $ref: '#/components/schemas/Chat'
      example:
        - chatId: 1
          userAvatar: url
          userName: name
          isChatBlocked: false
          isAuthUserBlockChat: false
          lastMessage: message
          lastMessageTime: 15-10-21 19:00:00
          isAuthUserMessage: true
          isSeenMessage: true
          numberNotSeenMessages: 0
          lastMessageType: text
        - chatId: 2
          userAvatar: url
          userName: name
          isChatBlocked: false
          isAuthUserBlockChat: false
          lastMessage: message
          lastMessageTime: 15-10-21 19:00
          isAuthUserMessage: true
          isSeenMessage: false
          numberNotSeenMessages: 4
          lastMessageType: image
...
    Messages:
      type: object
      properties:
        chat:
          $ref: '#/components/schemas/CurrentChat'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          example:
            - message: message1
              messageTime: 15-10-21 19:00:00
              isAuthUserMessage: true
              messageId: 1
              isMessageSeen: true
            - message: message2
              messageTime: 15-10-21 19:01:00
              isAuthUserMessage: false
              messageId: 2
              isMessageSeen: true
...
    Error:
      required:
        - code
        - detail
        - title
      type: object
      properties:
        code:
          type: integer
          format: int32
        title:
          type: string
        detail:
          type: string
    Chat:
      required:
        - chatId
        - userAvatar
        - userName
        - isChatBlocked
        - lastMessage
        - lastMessageTime
        - lastMessageType
        - isAuthUserMessage
        - isSeenMessage
        - numberNotSeenMessages
        - isOnline
      type: object
      properties:
        chatId:
          type: integer
          format: int64
          description: "Для подгрузки сообщений в чате для запроса используется в качестве параметра
           и для прослушивания события отправки сообщения в данном чате,
            в названии события необходимо будет указывать id чата"
        userAvatar:
          type: string
          format: uri
        userName:
          type: string
        isOnline:
          type: boolean
        isChatBlocked:
          type: boolean
        isAuthUserBlockChat:
          type: boolean
        lastMessage:
          type: string
        lastMessageTime:
          type: string
          format: data-time
        lastMessageType:
          type: string
          description: Тип сообщения, например, картинка, текст и т.д
          enum:
            - text
            - image
        isAuthUserMessage:
          type: boolean
        isSeenMessage:
          type: boolean
        numberNotSeenMessages:
          type: integer
          format: int32
    CurrentChat:
      required:
        - chatId
        - userAvatar
        - userName
        - isChatBlocked
        - isOnline
      type: object
      properties:
        chatId:
          type: integer
          format: int64
          description: "Для подгрузки сообщений в чате для запроса используется в качестве параметра"
        userAvatar:
          type: string
          format: uri
        userName:
          type: string
        isOnline:
          type: boolean
        isChatBlocked:
          type: boolean
...
    Message:
      required:
        - message
        - messageTime
        - isAuthUserMessage
        - messageId
        - isMessageSeen
        - messageType
      type: object
      properties:
        message:
          type: string
        messageTime:
          type: string
          format: data-time
        isAuthUserMessage:
          type: boolean
        messageId:
          type: integer
          format: int64
        isMessageSeen:
          type: boolean
        messageType:
          type: string
          description: Тип сообщения, например, картинка, текст и т.д
          enum:
            - text
            - image
    SendMessage:
      required:
        - message
        - chatId
      type: object
      properties:
        message:
          type: string
        chatId:
          type: integer
          format: int64
          description: "Чат создается между двумя пользователями, по его айдишнику будет поиск пользователя, которому отправлено сообщение"
        messageType:
          type: string
          description: Тип сообщения, например, картинка, текст и т.д
          enum:
            - text
            - image
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT